DP-6846 background color added to the BoxWrapper
