+++ packages/asc-ui/src/components/Header/Header.stories.tsx
@@ -5,11 +5,15 @@ import BoxWrapper from '../../internals/Box/BoxWrapper'
 import { ThemeProvider } from '../../index'
 import Header from './Header'
 import ContentFiller from '../../internals/ContentFiller/ContentFiller'
+import HeaderTall from './HeaderTall'
+
+const outsideBackgoundColor = '#eee'
+const contentBackgrountColor = '#F5F5F5'
 
 const HeaderStory: React.FC<{}> = () => (
   <ThemeProvider themeName="default">
     <>
-      <BoxWrapper>
+      <BoxWrapper backgroundColor={outsideBackgoundColor}>
         <Header
           title="Data en informatie"
           homeLink="http://data.amsterdam.nl"
@@ -17,9 +21,40 @@ const HeaderStory: React.FC<{}> = () => (
           menu={<CenteredElement> M </CenteredElement>}
         />
       </BoxWrapper>
-      <ContentFiller backgroundColor="#eee" />
+      <ContentFiller backgroundColor={contentBackgrountColor} />
     </>
   </ThemeProvider>
 )
 
-storiesOf('Composed/Header', module).add('default state', () => <HeaderStory />)
+const HeaderTallStory: React.FC<{}> = () => (
+  <ThemeProvider themeName="default">
+    <>
+      <BoxWrapper backgroundColor={outsideBackgoundColor}>
+        <HeaderTall
+          title="Data en informatie"
+          homeLink="http://data.amsterdam.nl"
+          search={<CenteredElement> Search </CenteredElement>}
+          menu={<CenteredElement> M </CenteredElement>}
+        />
+        <Header
+          title="Data en informatie"
+          homeLink="http://data.amsterdam.nl"
+          search={<CenteredElement> Search </CenteredElement>}
+          menu={<CenteredElement> M </CenteredElement>}
+          headerSize="tall"
+        />
+      </BoxWrapper>
+      <BoxWrapper backgroundColor={outsideBackgoundColor}>
+        <ContentFiller
+          backgroundColor={contentBackgrountColor}
+          maxWidth="1800px"
+          headerSize="tall"
+        />
+      </BoxWrapper>
+    </>
+  </ThemeProvider>
+)
+
+storiesOf('Composed/Header', module)
+  .add('Header short', () => <HeaderStory />)
+  .add('Header tall', () => <HeaderTallStory />)
